fix(flake): check if ~/.config/mbnvim exists before linking

+++ flake.nix
@@ -26,7 +26,9 @@
           ];
 
           shellHook = ''
-            ln -s $(pwd) ~/.config/mbnvim
+            if ! test -f ~/.config/mbnvim; then
+              ln -s $(pwd) ~/.config/mbnvim
+            fi
             export NVIM_APPNAME=mbnvim
           '';
         };
